Remove SID and remove SID hack
Since the mappings are defined in a function and they call functions
defined in this script, we're storing the SID value so we can bypass the
name mangling. Remove all of that.

I think we can do this better using <Plug>. Users would be able to
define mappings, but that's not so bad.

For now I'm removing it all so I can move stuff to autoload before
bringing it back (if necessary).